🔧 #1 페이지 구성 변경
